inference/requirements-openrouter.txt
# Minimal dependencies for DeepResearch with OpenRouter
# No local model deployment needed!

# Core - OpenAI SDK for OpenRouter API
openai>=1.0.0

# Agent Framework
qwen-agent>=0.0.26

# HTTP & Web Requests
requests>=2.32.0
httpx>=0.28.0
aiohttp>=3.9.0

# CLI & Environment
rich>=14.0.0
python-dotenv>=1.0.0

# JSON Processing
json5>=0.9.0
jsonlines>=4.0.0

# Data Validation
pydantic>=2.0.0

# Token Counting (for web content truncation)
tiktoken>=0.6.0

# Code Execution (PythonInterpreter tool)
sandbox-fusion>=0.3.0

# File Parsing
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.11.0
pdfminer.six>=20221105
python-pptx>=1.0.0
pillow>=10.0.0
pypdfium2>=4.0.0

# Utilities
tqdm>=4.65.0
tenacity>=8.0.0
python-dateutil>=2.8.0

# Optional: For Alibaba Document Parser (IDP)
# alibabacloud-docmind-api20220711>=1.4.0
# alibabacloud-tea-openapi>=0.4.0
# alibabacloud-credentials>=1.0.0

# Optional: For video analysis
# opencv-python-headless>=4.0.0
# soundfile>=0.12.0

